% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_select.R
\name{fm.select}
\alias{fm.select}
\alias{fm.select.FM}
\title{Select Best Model}
\usage{
fm.select(object, trace = NULL, best.iter = NULL, drop.trace = FALSE)
}
\arguments{
\item{object}{a FM object created by `fm.train`, which must contain \strong{Trace}
element, that's to say, \strong{step_size} should be greater than 0 in \code{track.control}}

\item{trace}{a FMTrace object}

\item{best.iter}{an integer, iteration number of trace.if \strong{best.iter} is missing
it will select the best model automatically}

\item{drop.trace}{whether to delete \strong{Trace} from model}
}
\description{
select the best model according to the performance on test data
}

