% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_track.R
\name{fm.track}
\alias{fm.track}
\alias{fm.track.FM}
\title{Track Model's Performance}
\usage{
fm.track(object, data = NULL, newdata = NULL, data.normalize = TRUE, newdata.normalize = TRUE,
                evaluate.metric = c("LL", "AUC", "ACC", "RMSE", "MAE"), max_threads = 1L)
}
\arguments{
\item{object}{a FM object}

\item{data, newdata}{objects created by fm.matrix}

\item{data.normalize}{whether to normalize data}

\item{newdata.normalize}{whether to normalize newdata}

\item{evaluate.metric}{a character, "RMSE"、"MAE" for regression model or
"AUC"、"ACC"、"LL" for classification model}
}
\description{
track model's performance on two datasets, the performance are
measured by \strong{evaluate.metric}
}

