% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_update.R
\name{fm.update}
\alias{fm.update}
\title{Update Factorization Machines Model}
\usage{
fm.update(object, data, normalize = TRUE, control = NULL)
}
\arguments{
\item{object}{a FM object, created by \link{fm.train}}

\item{data}{a fm.matrix object, created by \link{fm.matrix}}

\item{normalize}{whether to normalize data}

\item{control}{list of congtrols, \link{model.control}, \link{solver.control}, \link{track.control}}
}
\description{

}

