% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_solver_control.R
\name{FTRL.solver}
\alias{FTRL.solver}
\title{FTRL solver}
\usage{
FTRL.solver(...)
}
\seealso{
Other solver: \code{\link{ALS.solver}},
  \code{\link{MCMC.solver}}, \code{\link{SGD.solver}},
  \code{\link{TDAP.solver}}
}

