% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_solver_control.R
\name{MCMC.solver}
\alias{MCMC.solver}
\title{MCMC solver}
\usage{
MCMC.solver(...)
}
\seealso{
Other solver: \code{\link{ALS.solver}},
  \code{\link{FTRL.solver}}, \code{\link{SGD.solver}},
  \code{\link{TDAP.solver}}
}

