% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_predict.R
\name{predict.FM}
\alias{predict.FM}
\title{Predict Method for FM Fits}
\usage{
predict(object, newdata = NULL, max_threads = 1)
}
\arguments{
\item{object}{a fitted object of class inheriting from "fm".}

\item{newdata}{a data frame in which to look for variables with which to predict.}

\item{max_threads}{an integer, number of threads to speed up computing, \strong{openmp} should be supported.}
}
\description{
Obtains predictions from a fitted factorization machines model object.
}

