% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_solver_control.R
\name{SGD.solver}
\alias{SGD.solver}
\title{SGD solver}
\usage{
SGD.solver(...)
}
\seealso{
Other solver: \code{\link{ALS.solver}},
  \code{\link{FTRL.solver}}, \code{\link{MCMC.solver}},
  \code{\link{TDAP.solver}}
}

