% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_control.R
\name{solver.control}
\alias{solver.control}
\title{Auxiliary for controlling optimization routines}
\usage{
solver.control(max_iter = 10000, nthreads = 1, convergence = 1e-4, evaluate.method = "LL", solver = TDAP.solver())
}
\arguments{
\item{max_iter}{integer giving maximal number of iterations, 25 for MCMC/ALS and 10000 for other optimization routines by default}

\item{nthreads}{integer, number of threads to speed up computing, \strong{openmp} should be supported}

\item{convergence, evaluate.method}{control how to early stop}

\item{solver}{function to set parameters of optimization routine}
}
\description{
Auxiliary functions for optimization routines
}
\details{
SGD/FTRL/TDAP updates model every single example, while MCMC/ALS needs to sweeps through all the data for each update.
So the maximal number of iterations of MCMC/ALS is much smaller, which is 25 by default.

By now, only MCMC and ALS support parallel computing. And if openmp is not supported, nthreads makes no sense.

Early stopping is controlled by \strong{evaludate.method} and \strong{convergence}. If the difference of the value between two interations, which is
calculated by \code{evaluate.method}, is smaller than \code{convergence} for three times in a row, iteration will stop early.
\code{evaluate.method} includes:
\itemize{
\item "LL":loglikehood
\item "AUC": auc
\item "ACC": accurancy
\item "RMSE": root mean square err
\item "MAE": mean absolute error
}
"LL","AUC","ACC" are for CLASSIFICATION task. And "RMSE","MAE" are for REGRESSION task.

\strong{solver} is a function to set parameters of optimization routine further, including \link{ALS.solver}
\link{MCMC.solver} \link{SGD.solver} \link{TDAP.solver} \link{FTRL.solver}
}

