% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_solver_control.R
\name{TDAP.solver}
\alias{TDAP.solver}
\title{TDAP solver}
\usage{
TDAP.solver(...)
}
\seealso{
Other solver: \code{\link{ALS.solver}},
  \code{\link{FTRL.solver}}, \code{\link{MCMC.solver}},
  \code{\link{SGD.solver}}
}

