% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_track_control.R
\name{track.control}
\alias{track.control}
\title{Auxiliary for controlling tracking routines}
\usage{
track.control(step_size = -1, evaluate.metric = "LL", convergence = 1e-04)
}
\arguments{
\item{step_size}{record parameters every \strong{step_size} iteration}

\item{convergence, evaluate.metric}{control how to early stop}
}
\details{
Early stopping is controlled by \strong{evaludate.metric} and \strong{convergence}. If the difference of the value between two interations, which is
calculated by \code{evaluate.metric}, is smaller than \code{convergence} for three times in a row, iteration will stop early.
\code{evaluate.metric} includes:
\itemize{
\item "LL":loglikehood
\item "AUC": auc
\item "ACC": accurancy
\item "RMSE": root mean square err
\item "MAE": mean absolute error
}
"LL","AUC","ACC" are for CLASSIFICATION task. And "RMSE","MAE" are for REGRESSION task.
}

