% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_set_threads.R
\name{fm.set_threads}
\alias{fm.set_threads}
\title{Set Number of Threads}
\usage{
fm.set_threads(max.threads = 1)
}
\arguments{
\item{max.threads}{an integer, number of threads to speed up computing, \strong{openmp} should be supported.}
}
\description{
set the maximum number of available threads for parallel execution
}

