% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_solver_control.R
\name{solver.control}
\alias{solver.control}
\title{Auxiliary for controlling optimization routines}
\usage{
solver.control(max_iter = 10000, convergence = 1e-4, evaluate.metric = "LL", solver = TDAP.solver())
}
\arguments{
\item{max_iter}{integer giving maximal number of iterations, 25 for MCMC/ALS and 10000 for other optimization routines by default}

\item{solver}{function to set parameters of optimization routine}
}
\description{
Auxiliary functions for optimization routines
}
\details{
SGD/FTRL/TDAP updates model every single example, while MCMC/ALS needs to sweeps through all the data for each update.
So the maximal number of iterations of MCMC/ALS is much smaller, which is 25 by default.

By now, only MCMC and ALS support parallel computing.

\strong{solver} is a function to set parameters of optimization routine further, including \link{ALS.solver}
\link{MCMC.solver} \link{SGD.solver} \link{TDAP.solver} \link{FTRL.solver}
}

